refactor(Numbers): 클래스명, 메서드명 변경
